pyproject.toml
[tool.poetry]
name = "tachidesk-notifier"
version = "0.1.0"
description = "A script to track Tachidesk manga updates and send notifications"
authors = ["timesfai <timesfai@fastmail.com>"]
packages = [{include = "tachidesk_notifier"}]

[tool.poetry.dependencies]
python = "^3.9"
requests = "^2.26.0"
pyTelegramBotAPI = "^4.2.0"
python-dotenv = "^0.19.0"

[tool.poetry.dev-dependencies]
pytest = "^7.3.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
testpaths = ["tests"]